requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.15,<4.0
Pillow>=10.4,<12.0
reportlab>=4.2,<5.0
